Implement all current functionality with database

+++ backend/sim/tournament/TrainerStats.hpp
@@ -10,7 +10,7 @@ public:
     int losses = 0;
     int opponentRatingsTotal = 0;
     //Index into the results list
-    int bestWin = -1;
+    ssize_t bestWin = -1;
     int bestWinEloDiff = INT_MIN;
 
     void calculateElo();
